add movie and genre domain models and exposed entities, refactor database schemas
- Add `Movie` and `Genre` domain data classes
- Add Exposed ORM tables and entities for `MoviesTable`, `GenresTable`, and `MovieGenreTable`
- Update database migration scripts (V1, V2, V4, V6-V9) to change ID column types from `BIGSERIAL`/`BIGINT` to `BIGINT`/`INT`
- Rename `movie_genres` table to `movie_genre` in V6 migration
- Change `name` column type from `TEXT` to `VARCHAR(50)` in V4 `genres` table migration

+++ src/main/resources/db/migration/V7__rooms_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS rooms(
-    id BIGINT PRIMARY KEY,
+    id INT PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     rows INT NOT NULL,
     cols INT NOT NULL
