+++ src/main/resources/db/migration/V9__showtimes_table.sql
@@ -3,9 +3,9 @@ CREATE EXTENSION IF NOT EXISTS btree_gist;
 CREATE TYPE showtime_status AS ENUM ('SCHEDULED', 'CANCELLED');
 
 CREATE TABLE IF NOT EXISTS show_times(
-    id BIGINT PRIMARY KEY,
+    id INT PRIMARY KEY,
     movie_id BIGINT NOT NULL,
-    room_id BIGINT NOT NULL,
+    room_id INT NOT NULL,
     starts_at timestamptz NOT NULL,
     ends_at timestamptz NOT NULL,
     price BIGINT NOT NULL,
